requirements.txt
Flask==2.2.2
matplotlib==3.5.0
numpy==1.23.5
openai==0.28.1
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.8.1.78
paddleocr==2.7.0.3
pandas==1.5.3
paddlepaddle
pypdfium2==4.18.0
python_dateutil==2.8.2
requests==2.28.1
Werkzeug==2.2.2
